update not working

+++ src/main/java/com/appointment/repositories/AppointmentRepository.java
@@ -2,10 +2,9 @@ package com.appointment.repositories;
 
 import com.appointment.models.Appointment;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AppointmentRepository extends JpaRepository<Appointment, Long> {
+public interface AppointmentRepository extends JpaRepository<Appointment, Integer> {
 
 }
